mltr/requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9
Django==3.0.5
django-nose==1.4.6
Flask==1.0.2
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
newsapi-python==0.2.6
nose==1.3.7
numpy==1.18.2
opencv-python==4.2.0.34
paramiko==2.7.1
Pillow==7.1.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.1
sshtunnel==0.1.5
urllib3==1.25.8
Werkzeug==1.0.0
wincertstore==0.2
